separate JsonLog
