+++ LuckParser/LuckParser.csproj
@@ -121,6 +121,7 @@
     <Compile Include="Models\BossLogic\Sabetha.cs" />
     <Compile Include="Models\BossLogic\Gorseval.cs" />
     <Compile Include="Models\BossLogic\ValeGuardian.cs" />
+    <Compile Include="Models\DataModels\JsonLog.cs" />
     <Compile Include="Models\DataModels\ParseEnum.cs" />
     <Compile Include="Models\DataModels\ParsedLog.cs" />
     <Compile Include="Models\DataModels\SettingsContainer.cs" />
